# Speed up Point3D.X/Y/Z get/set