This PR improves the average time of the following loop by around **10%**.

Here is a small benchmark I wrote to accompany my above findings: https://github.com/EoD/EDDBenchmarks/blob/bc025004ea492b2cf01343baa30e2a8c1d0960fc/EDDBenchmarks/Point3DDistanceBenchmark.cs#L47-L70
Original:
```
Generating random Point3D list of size 25000
List generated, duration: 0.003s

Calculating distances
Total: 3.381957E+07, duration: 6.334s
Total: 3.381957E+07, duration: 6.198s
Total: 3.381957E+07, duration: 6.194s
Total: 3.381957E+07, duration: 6.185s
Total: 3.381957E+07, duration: 6.210s
```

With the patch around 10% faster
```
Generating random Point3D list of size 25000
List generated, duration: 0.002s

Calculating distances
Total: 3.38295E+07, duration: 5.646s
Total: 3.38295E+07, duration: 5.616s
Total: 3.38295E+07, duration: 5.614s
Total: 3.38295E+07, duration: 5.614s
Total: 3.38295E+07, duration: 5.641s
```

EDIT:
There is also a major performance difference with debugging enabled. The difference is **>400%**. 
These are screenshots from the dotTrace Profiler:

Original:
![dotTrace_before](https://user-images.githubusercontent.com/293499/67119858-4d1d8e00-f1e8-11e9-978b-48a82e5fc0f7.png)


After the changes:
![dotTrace_after](https://user-images.githubusercontent.com/293499/67119781-1e071c80-f1e8-11e9-9d70-727bb3d9700d.png)