@robbyxp1 2019-10-20
I'm surprised this did anything. Did you run it in release mode to see if there was a difference? 

Also your making the setters private, thus changing the interface..