@iainross 2019-10-21
foo { get { return bar; } } and foo { get => bar; } are equivalent, the latter is just a bit of syntactic sugar to save us some key strokes, the IL is meant to be identical.