@EoD 2019-10-21
@iainross yeah, I was just trying to find why it makes a performance difference in the Debug build.
A random guess: internally the new syntax is not a function call and this one function call causes some overhead in the debug build?